Move copyToClipboard util into @uppy/dashboard.
