+++ packages/@uppy/dashboard/src/copyToClipboard.js
