+++ packages/@uppy/dashboard/src/copyToClipboard.test.js
